frontend/requirements.txt
protobuf~=5.29.1
grpcio~=1.68.1
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
pip~=23.2.1
wheel~=0.41.2
MarkupSafe~=3.0.2
werkzeug~=3.1.3
click~=8.1.7
jinja2~=3.1.4
blinker~=1.9.0
itsdangerous~=2.2.0
visitor~=0.1.3
setuptools~=68.2.0
importlib-metadata~=8.5.0
zipp~=3.21.0
kafka~=1.3.5
kafka-python~=2.0.2